+++ classes/prefix_WPgutenberg/blocks/accordion-item/attributes.js
@@ -1,0 +1,8 @@
+const attributes = {
+  accordionTitle: {
+    type: 'string',
+    default: 'Label',
+  }
+}
+
+export default attributes;