+++ classes/prefix_WPgutenberg/blocks/accordion-item/inspector.js
@@ -1,0 +1,58 @@
+/**
+ * Internal block libraries
+*/
+const { __ } = wp.i18n;
+const { Component } = wp.element;
+const { registerBlockType } = wp.blocks;
+const {
+  ColorPalette,
+  PanelColorSettings,
+  ContrastChecker
+} = wp.editor;
+
+const {
+  InspectorControls
+} = wp.blockEditor;
+
+const {
+  CheckboxControl,
+  PanelBody,
+  PanelRow,
+  RadioControl,
+  RangeControl,
+  TextControl,
+  TextareaControl,
+  ToggleControl,
+  SelectControl
+} = wp.components;
+
+/**
+ * Create an Inspector Controls wrapper Component
+*/
+export default class Inspector extends Component {
+  constructor() {
+    super(...arguments);
+  }
+  render() {
+    const {
+      attributes: {
+        accordionTitle
+      },
+      setAttributes
+    } = this.props;
+
+    return (
+      <InspectorControls>
+          <PanelBody title={ __( 'Accordion item settings', 'WPgutenberg' ) } >
+            <PanelRow>
+              <TextControl
+                label={__("Label", "WPgutenberg")}
+                value={accordionTitle}
+                onChange={accordionTitle => setAttributes({ accordionTitle })}
+              />
+            </PanelRow>
+          </PanelBody>
+      </InspectorControls>
+    );
+  }
+}