+++ classes/prefix_WPgutenberg/blocks/index.js
@@ -4,3 +4,5 @@
 import "./vimeo";
 import "./posts";
 import "./postsfilter";
+import "./accordion";
+import "./accordion-item";